website/requirement_dev.txt
Django==1.8
djangorestframework==3.1.1
MySQL-python==1.2.5
Pillow==2.8.1
uWSGI==2.0.10
